Move main.py to scripts/nwp_data as requested by maintainer
