+++ pyproject.toml
@@ -66,7 +66,7 @@ dev = [
 ]
 
 [project.scripts]
-open-data-pvnet = "open_data_pvnet.main:main"
+open-data-pvnet = "open_data_pvnet.scripts.nwp_data.main:main"
 
 [tool.black]
 line-length = 100
